app/api/requirements.txt
Flask==3.0.3
flask_cors
python-dotenv
markitdown
openai
